Making repository as interface to be able to mock and easy to unit test it
